backend/requirements.txt
flask==2.2.3
psycopg2-binary==2.9.5
flask-cors==3.0.10
# FIJAR VERSIONES DE WERKZEUG Y JINJA2 PARA EVITAR INCOMPATIBILIDAD
Werkzeug==2.2.3 
Jinja2==3.1.2